django_app/requirements.txt
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv
